Organ pane relayout ready

+++ src/View/Panes/ChannelsSetupPane.h
@@ -10,8 +10,8 @@ class View;
 class ChannelsSetupPane : public Pane
 {
  public:
-   ChannelsSetupPane(View &view, MixerSubModel &mixerSubModel, int leftPercentage, int topPercentage,
-    int widthPercentage, int heightPercentage);
+   ChannelsSetupPane(View &view, MixerSubModel &mixerSubModel, double leftPercentage, double topPercentage,
+    double widthPercentage, double heightPercentage);
 
    void Fill() override;
    void Relayout() override;
