+++ src/View/Panes/OrganPane.h
@@ -12,8 +12,9 @@ class OrganPane : public Pane
  public:
    static const int NR_OF_DRAWBAR_SLIDERS = 9;
 
-   OrganPane(View &view, OrganSubModel &organSubModel, int leftPercentage, int topPercentage, int widthPercentage,
-    int heightPercentage);
+   OrganPane(View &view, OrganSubModel &organSubModel, double leftPercentage, double topPercentage,
+    double widthPercentage,
+    double heightPercentage);
 
    void Fill() override;
    void Relayout() override;
